+++ src/org/soulspace/overarch/adapter/render/plantuml/uml.clj
@@ -371,7 +371,8 @@
 (defn render-uml-layout
   "Renders the layout for the UML diagram."
   [view]
-  (let [spec (:spec view)]
+  (let [spec (:spec view)
+        plantuml-spec (:plantuml spec)]
     (flatten [;(when (:styles spec)
               ;  (into [] (map #(render-uml-style view %)) (:styles spec)))
               (when (:sketch spec)
@@ -381,7 +382,11 @@
               (when (:layout spec)
                 (uml-layouts (:layout spec)))
               (when (:linetype spec)
-                (puml/linetypes (:linetype spec)))])))
+                (puml/linetypes (:linetype spec)))
+              (when (:node-separation plantuml-spec)
+                (str "skinparam nodesep " (:node-separation plantuml-spec)))
+              (when (:rank-separation plantuml-spec)
+                (str "skinparam ranksep " (:rank-separation plantuml-spec)))])))
 
 (defmethod puml/render-plantuml-view :uml-view
   [model options view]